+++ client/gutenberg/extensions/map/editor.js
@@ -3,31 +3,7 @@
 /**
  * Internal dependencies
  */
-import { settings as mapSettings } from './settings.js';
+import * as block from './index.js';
 import registerJetpackBlock from 'gutenberg/extensions/presets/jetpack/utils/register-jetpack-block';
-import edit from './edit';
-import save from './save';
-import './style.scss';
-import './editor.scss';
 
-export const { name } = mapSettings;
-
-export const settings = {
-	title: mapSettings.title,
-	icon: mapSettings.icon,
-	category: mapSettings.category,
-	keywords: mapSettings.keywords,
-	description: mapSettings.description,
-	attributes: mapSettings.attributes,
-	supports: mapSettings.supports,
-	getEditWrapperProps( attributes ) {
-		const { align } = attributes;
-		if ( -1 !== mapSettings.validAlignments.indexOf( align ) ) {
-			return { 'data-align': align };
-		}
-	},
-	edit,
-	save,
-};
-
-registerJetpackBlock( name, settings );
+registerJetpackBlock( block.name, block.settings );
