+++ client/gutenberg/extensions/presets/jetpack/editor.js
@@ -10,7 +10,7 @@ import './editor-shared/block-category'; // Register the Jetpack category
 // Appending `Block` to the names to keep `Map` from colliding with JS' Map
 import * as ContactFormBlock from 'gutenberg/extensions/contact-form/editor';
 import * as MarkdownBlock from 'gutenberg/extensions/markdown/editor';
-import * as MapBlock from 'gutenberg/extensions/map/editor';
+import * as MapBlock from 'gutenberg/extensions/map';
 import * as PublicizeBlock from 'gutenberg/extensions/publicize/editor';
 import * as SimplePaymentsBlock from 'gutenberg/extensions/simple-payments/editor';
